server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiomysql==0.2.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyjwt==2.8.0
pymysql==1.1.0
DBUtils==3.1.0
tushare==1.4.4
pandas==2.1.4
numpy==1.26.2
sqlalchemy==2.0.23
schedule==1.2.0
httpx==0.25.2
volcengine_python_sdk==5.0.7
